Rename setupFixture to setup-fixture
